Add support to run RemoteStoreIT with S3 integration
